move to_latex back to latex_generation.py
