+++ pyquil/latex/__init__.py
@@ -1,3 +1,3 @@
-from pyquil.latex.latex_generation import DiagramSettings
-from pyquil.latex._diagram import to_latex
+from pyquil.latex.latex_generation import to_latex
+from pyquil.latex._diagram import DiagramSettings
 from pyquil.latex._ipython import display
